requirements.txt
streamlit
speedtest-cli